Spider script for programmatically pulling in contract configs directly from blockchain (#8)
This is an implementation of the Spider method to programmatically pull contract configs directly from the blockchain. The script pulls contract names, addresses, ABIs, etc., directly from Etherscan using a Saddle import Hardhat plugin. A summary of config management using the Spider method is described in RFC 027. 

More details can be found in the PR's description as well.