Workers and Loaders
This patch starts to build a workers/loaders framework that can run multiple scenarios, in parallel. We also add globbing to find scenarios (that will be set by config, once that's set-up). We will work to allow scenarios to be in JavaScript or Typescript, with probably a preference towards typescript.