Improve Strategy in re: HardhatContext and Scenario Context
This patch helps accomplish two goals: first, it makes it where the Hardhat that's constructed in each worker more correctly mirrors the hardhat context that was created in the parent. It's not a perfect match, per se, but we copy enough data and/or code paths to make it close enough to pass most issues we might otherwise see.

Secondly, we make `getInitialContext` and `forkContext` a per scenario set of code, for now. We make a project-code based helper that makes it easy to reuse this code, and thus free ourselves of constantly fighting for global coherence (since we can only prove per-test coherence anyway). As such, the code becomes simpler and we have less type issues, plus the code now is running properly.