Start to add idempotent deploys
This patch starts to add idempotent deploys to the initial context, that should deploy the raffle contracts for a given network if they are not discovered by the spider script. This should be mostly working at this point, though it needs to be set-up for real networks (e.g. development and goerli) and I'm currently getting a revert trying to read data from the raffle contract.