Add Remote Token Constraint (#60)
* Add Remote Token Constraint

This patch adds a remote token constraint, which pulls in a token from a random network (e.g. mainnet) and deploys it to whatever base the scenario is currently running in (e.g. development or goerli). This allows us to test against real tokens without needing to copy-and-paste the source code or build files manually. E.g. we can test supporting new markets not _just_ on mainnet.

There's lots of little tweaks in this branch, e.g. to ease the type signature of `solve` since null can mean "don't do anything" and/or returning a single solution can mean "just use this one solution." Finally, that solution doesn't need to return a value if it decides, instead, to mutate context (versus returning something new).

Overall, this is just an attempt to start to get more complex constraints and scenarios built to start to flex the framework and see how things are shaping up.

* Address basic feedback; allow no constraints; use concat for