Support aliases for contracts in `relations.json` (#53)
This PR adds an optional rule to relations.json called alias. This rule is used to derive the key that is assigned to a contract in pointers.json. If this rule is not provided, the contract name will be used as the alias instead. This rule has two special characters: @ and +.

@ followed by a function name is used to read a value from that contract's function. + is used as a delimiter.

Example: @symbol+Delegator will equate to cDaiDelegator for cDai's delegator contract.

Other changes in this PR:

* Move alias logic from standalone Spider script to DeploymentManager
* Add a new type to DeploymentManager called AliasesMap to keep track of all the aliases for an address
* Delete standalone Spider script
* Support writeToDisk option for Spider that writes a pointers.json file to cache