Add Base Comet Scenarios
This patch adds some base comet scenario code. We work to unify development deployments with the development script, itself, which is called when running scenarios on development (but not for other networks). More WIP.

Start to get deploy scripts and development flow working again

WIP: Run Spider on Development (#88)

* WIP: Run Spider on Development

This patch adds support for effectively running spider on development. We increase support for running Deployment Manager without hitting disk (cache is just stored in memory), and for development, we make sure that all deploys write the information that would have been spidered. Thus, when we run "spider" on development, it should have all of the necessarily build artifacts already loaded. This should hopefully help unify how we look at deployment management, such that development is not full of one-off deviations. We're currently fighting getting the hardhat build artifacts to align with traditional solc artifacts, but overall, this patch should be close to working. The idea in a nutshell is:

* If you deploy from `npx hardhat node && yarn deploy --network development`
* Or: `yarn deploy --network goerli`
* Or: `yarn sceanrio --bases development`
* Or: `yarn sceanrio --bases goerli`

In each of those cases, we should always be getting and pulling configuration from the spider metadata. In the cases of development, it will be put there after running `scripts/deploy-comet.ts` (to disk for a hardhat node and just in memory for a development scenario). In the cases of goerli, it should come from remote. There's some middle ground that will need to be decided (e.g. deployments in a scenario to goerli should also pull from memory).

So, all-in-all, this patch is attempting to unify the many ways to get deployment data, via spider, in cases where things might truly come from Etherscan or otherwise in cases where we push to that cache ourselves on deploy.

* Minor PR clean-up

* Fix running spider on local scenarios

* Fix file location

* Deploy fresh Goerli

* Deploy Fuji