Rename comet.ts to deploy.ts
This patch simply renames `comet.ts` to `deploy.ts`, since that's its primary use-case, and it makes it easy to search files for `deploy.ts`. This could also be `deploy-comet.ts`, but I think we have a script named that, so I don't want them to be confused. This file (`src/deploy.ts`) is going to be the primary "build a new comet deployment" for a while, and will probably grow in importance. It will eventually splinter into a "deploy versus migrate" concept, but it doesn't, per se, make sense to do that yet. Note: deploying a faucet token should one day be gated by network, as well.