Fix Scenario
This patch fixes scenarios by adding back in contracts we use in deployments. Additionally, we improve the speed and efficiency of scenarios by compiling and spidering before calling into scenarios, which means scenarios need to do less spidering themselves (which was being done inefficiently). This should fix the majority of problems we see in CI scenario runs. We also cache the deployment (spider) artifacts, so that spiders can run faster in the future.