Setup Modern Constraint
This adds a constraint (currently called `ModernConstraint`) that ensures that for a given scenario, the current network has an up-to-date Comet contract that matches the current code. This is accomplished by setting a requirement `{upgrade: true}` and then the constraint will perform an upgrade on the TransparentUpgradeableProxy contract. If this is not specified, then you will be using the code from current chain (which may be out of date with your changes). This is the start in the "how do I get to a state I need to test my changes"-world, but it does not attempt to choose a long-term solution, just one that is valuable for testing out our new code developments.