Run Spider Before Scenarios
This patch adds a subtask to the `scenario` task to run spider before running scenarios. This is very important since scenarios will run spider non-destructively, and if there's a lot that needs to be pulled from Etherscan, it ends up running those pulls _once per scenario_. We can consider other approaches of using shared/cached spider results, but this solves the 99% use-case problem, and itself should never be harmful. It _could_ slow down running scenarios a bit if the spider wasn't necessary, -- but scenarios run `dm.spider()` anyway, so probably not as much as you might think.