Add interest rate functions, getUtilization, and getPresentValue (#116)
This PR adds support for interest rate related functions. Interest rate configuration constants are passed to the constructor as per year rates and are converted/stored as per second rates by the constructor. This way, it is easier to manage the deploys/upgrades since annual rates are more readable than per second rates. However, this does mean the contract makes an assumption on the number of seconds in a year.