Add Utilization Constraint
This patch adds a utilization constraint. We have to figure out exactly how much `base` to supply or borrow (or both, if there was no starting supply) to reach a target utilization ratio. We then also need to get the relative prices and figure out how much collateral to supply to get there. We then need to source that collateral. Plus we need to actually supply or borrow, which also, is not currently in Comet. So lots of little pieces. We also:

 * Stub supply, withdrawal and initialize as they are needed to get a working test
 * Fix a bug in hardhat config for `gasPrice !== gas`
 * Add optional logging for development
 * Add optional `--sync` option for workers that do not use threads (useful for extended debugging messages)
 * Remove unused contract (npm) dependency
 * Switch from `Token.sol` to `ERC20.sol` which is more canonical
 * Fix a malfunctioning `erc20.json` dependency
 * Fix (mostly) `sourceToken` that had an infinite stack bug
 * Add sourcing of tokens from known accounts (e.g. the creator) for dev. Many tokens don't emit `Transfer` in the constructor.
 * Add option for `upgradeAndCall` on `upgradeTo`
 * Reset assets in a redeploy (this is really wrong but needs to be thought of in the bigger network context).
 * Add `allocateActor` that creates a clean account with some Eth to prevent "reuse" errors on known accounts.