Test-net Configuration
This patch is an attempt at setting test-net configuration via a simple `configuration.json` file for each network to make it easy to set configuration parameters, tokens, etc, that we would like for each network.

The alternative is to just try to do some "migrations" or "migration scripts" code, though I'm not entirely sure this can't be an aid in that. In part, this all depends on how Configurator works.

Also, here we try to "clone" tokens from the associated "mainnet" (e.g. Avalanche tokens to Fuji), as that's a common desire for our test-nets.