Fuzz requirement instead of comet config
Fix Issues with Forking and Snapshotting

This patch fixes issues with forking context and snapshotting. We ensure that forks actually do not conflict with one another by deep-cloning any necessary functions. Secondly, we improve snapshots so they are used between solutions of a solution-set. We probably need to work on what destructuring from CometContext looks like, which will be added to this branch before it's merged.

Context Properties

This patch creates "ContextProperties", which is a transformed version of context that is passed into scenarios. This allows us to not have to attach every possible value to the context object for destructuring, which becomes an issue when an object like `comet` changes (and thus should be received by an async getter), which either makes the scenario code more complicated (`let comet = await context.getComet()`), or we say `context.comet = comet` and then later if that value changes, we need to effectively observe those changes and reset this mirror-object. Instead, now, we create a function that is effectively: `getContextProperties(context) -> { comet: ..., ... }`. This function then can pull the data from the canonical source, but now the scenario can just destructure `{comet}` from this and not have to deal with guessing if this value is up to date (or awaiting it).

Additionally, we change `cache` to use `Map`s instead of objects for tracking its data and perform a smarter version of `deepClone`, which previously had a mistake that is was transforming `Map`s and other non-JSON data structures into flat objects, which then was failing further code. The new code deep clones Maps correctly and otherwise takes values as-is.