Refactor of Deployment Manager
Deployment Manager was born out of the idea that spider was crucial to the scenarios system. That is, after you spider, you should have access to then read those contracts. When you deploy contracts, they should be stored similar to how spider might crawl them. But this led to a concern that DeploymentManager was picking up a large amount of state that it handled inconsistently.

This patch reworks Deployment Manager such that it uses a single clean state object "cache" that is a read-through, write-though cache. That is, basically, it stores things in memory as it reads and writes from disk to make it fast, but if it didn't, it would work the same, just slower. This makes a convenient-use module that is used for scenarios and other tasks.

We refactored Deployment Manager sub-areas as modules with clean exported functions, which now have unit tests. This allows us to cleanly understand the behavior of say "spider" through a variety of configurations.

We refactor Relations.json to now be part of the TypeScript Hardhat config. This is 1) to gain the benefit of type-checking on the file, which was becoming more complex, and 2) to allow us to use small amounts of JavaScript for things like "how do I read the asset list from a Comet contract?."

We refactored Spider's use of aliases. Specifically, when looking for more contracts (the branch out "discovery" part of spider), we now mark those discovered contracts with the alias from the parent (e.g. Comet might say that its assets should be aliased by their ERC-20 symbol). We then allow the user to specify the configuration for those contracts via the alias or the contract name. So we could say that "DAI" has a special spider configuration and/or that "FiatTokenProxy" (USDC) does. This should allow us better diversity in handling different real-world situations, and is particularly well-suited for the current spidering of Comet.

We refactor Proxies to be `alias -> proxy address` from `address -> proxy address` since aliases.json already contains the link `alias -> address` and it should be eaiser for a human to read what this means.

Overall, this code should be significantly better and easier to use than the previous code.

Note: this is a WIP still as we need to finish fixing the interface of DeploymentManager and make sure that scenarios will now pass with these changes. These final changes should be very light compared to the overhaul already in this PR.

Finally, we deploy a new version of Comet to Fuji.