Restructure scenario runner and get it working for scenarios with multiple solutions (#123)
List of changes to different components of the scenario framework:

Workers/Runners:
1. Instead of each worker having a single runner that runs all bases for a scenario, a worker now creates a separate runner per base. Each runner holds a `World` for a different base. A single scenario will be run on all runners.
2. The runner takes a single scenario now instead of a list of scenarios.
3. The runner now returns the `Result` directly back to the worker instead of directly posting this to the parent. The worker then aggregates all the `Result`s from each runner and posts them to the parent. This is to prevent the first runner's `Result` from prematurely exiting the worker.

Snapshots: 
We had discussed only snapshotting after context initialization, but this is problematic because the context initialization is dependent on the scenario (e.g. await scenario.initializer(world)). Instead, I maintain two types of snapshots: a global one for the World and a local one for the context. The World snapshot is for resetting the runner back to a clean state of the world. The local context snapshot is for re-using the same context between multiple solutions of the same scenario.