Stall Telemetry (#158)
> Stall! Stall! Stall!

Not the best sound to hear when you're in the cockpit of a flight 30,000 feet above the majestic Indian Ocean, but music to the ears of a developer patiently waiting for Professor G.T. Hub to release your latest marks. Yes, yes, the office is closed today. There will be no marks. Well, at least it's better than wasting a calm afternoon sitting outside the professor's door, waiting.

P.S. We revert to base-scenario flattening, in hopes of better concurrency.