Kovan Migration
> Are you suggesting coconuts migrate?

This patch improves migrations, with the intent of building out a proper Kovan test-net.

First: we improve migrations by storing and retreiving the artifacts returned from `prepare` (and provide them to `enact`). This allows us to, say, deploy a bunch of contracts in the preparation step, store them, and later create a governance proposal that references those deployed artifacts. We aim to create a system that allows review prior to creating a governance proposal.

Second: we create a `gen:migration` task that will create a shell migration script for you. The primary purpose is we previously had migrations numbered sequentially, but that doesn't work if multiple PRs are creating migrations. So we use a timestamp-name combination now to prevent conflicts. This does not guarantee run-order, but nothing does, and that feels okay for now. E.g. some migrations might be merged into master but never run or accepted by governance. Note: there is currently no "run all migrations" command- each is run one-off, so skipping some is fine.

Third: We add `clone` to deployment manager (which is a light wrapper around import and deploy), which alleiviates a lot of the complexity in our previous migration script. We opt to use contract map overrides instead of "writing aliases," and overall, our migrations have become a lot simpler. We move Fuji's migration to our new style.

Fourth: We add `simulation` to the migrate task. This allows a user to see the effects of running a migration before running it actually. This should alleiviate a lot of concerns of "I wasted 50 ETH trying to test if my task worked." This will also allow others to test the effects of someone else's migration before it runs. This is not part of the scenario-framework (which will separate test migration scripts), but just a convenience for users of the CLI.

Fifth: we improve the `hardhat migrate` CLI commands to be easier and more intuitive, and we add a `MIGRATION.md` to explain the process for new users.

Sixth: we add a Kovan `configration.json` and migration. This will allow us to have a test-net on both Fuji and Kovan. Note: we opt into Kovan instead of Goerli as ChainLink doesn't have Goerli contracts, and we'd prefer to not have to simulate prices ourselves, even on test-net. We choose high-quality assets for the test-net, namely: USDC (base), WETH, WBTC, COMP, LINK and UNI.

Seventh: Brave Brave Sir Robin, Sir Robin Ran Away.

Eighth: We begin to attach to GitHub Actions as our runner. Specifically, the goal is that a user will run migration scripts in GitHub Actions (as opposed to "I ran this locally, here's my output"). This is to make it easier for auditors to understand the code that was run that created a given set of contracts and takes a risk out of the equation (e.g. a user says they deployed one thing, but actually deployed another). This isn't an end-all-be-all, but it should be helpful in general.

Ninth: We remove Goerli and Kovan from Bases, since they are temporarily non-functional until we get a proper Kovan deploy up, via the GitHub Actions above.