Add Kovan Deployment
This patch adds a Kovan deployment. Specifically, we take the results from the GitHub Action deployment here: https://github.com/compound-finance/comet/actions/runs/1822702784 and then attach them to `kovan/roots.json`. We had to cajole one contract to verify by manually performing the verification, but overall, the goal is to let the robots do the work to make it easier for humans to understand what happened, and thus this Kovan deployment is a great start, since it was deployed by GitHub.

Also note: we truly clone out the cache memory between runs now, since otherwise it was being potentially overwritten between runs.