Fix Issues with Forking and Snapshotting
This patch fixes issues with forking context and snapshotting. We ensure that forks actually do not conflict with one another by deep-cloning any necessary functions. Secondly, we improve snapshots so they are used between solutions of a solution-set. We probably need to work on what destructuring from CometContext looks like, which will be added to this branch before it's merged.