Add basic informational events
Currently added Supply, Transfer, and Withdraw events for both base and collateral.
This barely fits contract size using 1 optimizer run, and doesn't fit with 200 runs.
It would be 16 bytes cheaper to have only Supply/4 and Withdraw/4 versions of the events, but since we need Transfer/3 for ERC20 compatibility, I kept them all consistent.

If we want events to track collateral seized / debt repaid (in absorb), it'll be a bunch more expensive.
Likewise for when collateral is purchased from the protocol.