The 5 Asset Borrow Test [Not by Tom Wolfe]
This patch sets up a gas scenario to test the 5 asset borrow test, which is to compare a mainnet transaction with a hyper-similar test-net transaction to test what the expected gas would be for Comet. The goal is to take a transaction that is 200-300K gas on mainnet, and see if the similar one could be about 100K in Comet.

We libraryify a lot of things in this patch. This is just refactoring to allow us to re-use code. Things like `deployments/kovan/configuration.json` probably should be dropped in support of nicer libraries, etc, but this is just a start in that direction. For instance, our `deploy_kovan` migration now uses libraries so other code-spots can use "clone token" which was being done in that migration. We similarly share `upgradeComet` from ModernConstraint.

Initial reports are showing that the trx cost from v2 dropped from 254,584 gas to 143,351 gas, which represents a 43.70% reduction in gas costs.