Balance constraint + Transfer scenario (#235)
This PR adds support for specifying generic assets (e.g. "asset0") instead of a specific asset (e.g. "COMP") in the balance constraint. The issue with specifying a specific asset is that the asset may not exist for certain bases. This changes allows scenarios using the balance constraint to just work across all bases.

The balance constraint now looks something like this:

```
scenario(
  'Comet#balance constraint',
  {
    balances: {
      albert: { $asset0: 100 }, // in units of asset, not wei
    },
  },
```

Notice that the asset amount is specified in units of an asset instead of its wei amount. This is because asset0 may have different decimal scales on different bases. Using units instead of wei means scenarios will not have to make assumptions about the number of decimals for an asset.

This PR also includes an example of a transfer scenario using the balance constraint. More to be added in follow-up PRs.