Add Migration Constraint
This patch adds Migration Constraint, which runs each pending migration for each scenario to make it easy to automatically verify that each property tested by a scenario will continue to hold if that migration were to be prepared and enacted. A few notes:

 1. We currently are doing subsets, but we should considered ordered subsets, i.e. if migrations are run in different orders, do the properties still hold.
 2. Migrations currently are very bad at knowing if they were enacted. The hope is that this will be significantly easier when they start using Governance (e.g. they can just check to see if a Proposal with the same name/description/actions has been enacted. We could add the ability to check on-chain if a migration has been run, but that's costly and more complex than leaving it up to the migration itself to decide with the `enacted` function each migration must implement.
 3. Currently this is opt-in via `{ migrate: true }` flag. We should consider what requirements we want (e.g. is this just automatically enabled for all scenarios? Can scenarios specify they _should break_ if certain migrations are run?).

All-in-all, this should be a powerful tool to test that properties hold even after changes to the protocol are made, and has been a cornerstone of the intention of scenarios from inception.