Configurator (#160)
This is an implementation of the advanced architecture described in RFC 021, but with some changes to enhance upgradeability, namely a separate Configurator contract that is upgradeable. The overall system looks something like:

Timelock -> ProxyAdmin -> Proxy -> Configurator
                       \
                        > Proxy -> Comet

Timelock is an admin of ProxyAdmin, which is an admin of the two proxies. Timelock is also an admin of Configurator, meaning only the Timelock can set parameters in Configurator.