[ChainSecurity 5.10] Document risk of overflow in present value calculation (#316)
This PR discloses the risk that the present value of the int base balance will overflow at 2^103/2e18=~10 trillion for a base asset with 18 decimals. For USDC and other assets with less decimals, this hard cap is much higher.

If we were to use safe uint conversions in the present value functions, it would add ~700 gas to the supply/transfer/withdraw base functions. Since the hard cap is really high, we decided it's better to disclose the risk and keep the optimization.