Set-up Typescript Configuration (#233)
This patch properly sets up TypeScript configuration so that other projects can import this and use it correctly. We effectively make sure that the project is built when it's imported to another project. We make sure that only the `dist/` directory is included in npm and that it's picked up when used in a project (TypeScript is behind in its integration of package.json's `exports` key, so it's a little complex). This is in line with the requirements of `compound-js`'s integration of Comet.

Co-authored-by: Torrey Atcitty <torrey@compound.finance>