Accrue inside getReserves and add totalBorrow()
Based on feedback from integrating UI, also added notes to rate model fns which explicitly do *not* accrue.

Not sure whether it's better to drop the `accrueInternal` on `withdrawReserves` and `buyCollateral`.
If we do, it only saves 16 bytes contract size and a little gas, and probably the most audit-worthy change.