Derive store front discount from liquidationFactor (#327)
The store front discount is now derived from a collateral asset's liquidationFactor. The formula is:

DiscountFactor = StoreFrontPriceFactor * (FACTOR_SCALE - Asset.LiquidationFactor), where FACTOR_SCALE=1e18

The storeFrontPriceFactor has now been repurposed to be the fraction of the liquidation penalty (1 - liquidationFactor) that goes to the buyer of collateral instead of the protocol.

e.g. A storeFrontPriceFactor of 1e18=100% means the buyer's discount is equals to the entire liquidation penalty. A value of 0 means the buyer will have to buy the collateral at market price (without a discount).