Real testnet configurations (#338)
This also modifies some params in scenarios in order to work on testnet configurations with higher minimums and interest rates. The timing of some of the interest accrued seems excessive in some cases and is noted.

This also fixes a regression in how amounts are parsed, used to set exact collateral balances. We will probably want to add some more constraints for things like prices.