Add cached/cloned to DeploymentManager for idempotency in migrations
This adds a few fns to DeploymentManager that can be used in migration scripts instead of `deploy` (`cached`) and `clone` (`cloned`).
Using these, one can transform a valid migration script into a (partially) idempotent version which can reuse existing deployments at will.
This is primarily useful for debugging, and should not be used in migration scripts run in production, since searching artifacts by contract name is not a reliable process.