Add max repay functionality and remove getBorrowLiquidity (#373)
This PR adds the functionality for a user to repay all of their borrow position. 

Increases the contract size by 28 bytes. Now, the CometFactory contract is slightly over the contract limit (24.598kb vs limit of 24.576kb). Since the contract size is over the limit, we decided to remove getBorrowLiquidity() to reclaim some space. The resulting contract size is 24.364kb.