Add ability to sweep funds from Bulker (#413)
This PR adds an admin to the Bulker contract and allows the admin to sweep ERC20 tokens and ETH out of the contract to handle accidentally sent funds. This addresses part of #376's concern

This PR also removes the getAmount helper used in the Bulker to calculate the max transfer/withdraw values, as max transfer/withdraw are now supported natively in the protocol.