Fix CometActor/NonceManager forking (#435)
This PR fixes a bug where the wrong nonces were being used in a new scenario I was trying to write (the one created in this PR).

The root issue is that when a scenario context is forked, the CometActors in the forked context use the signers from the old deployment manager instead of the new deployment manager. This is a problem because we recently started using NonceManager to represent signers in deployment manager. NonceManagers are stateful and therefore can have their nonces get out-of-sync if different NonceManagers are used for the same signer. In the new scenario introduced in this PR, we use deployment manager to deploy a new contract directly in the scenario, which uses the signer from the new deployment manager. Then, when we try to fast execute a proposal, the old signer from the CometContext's CometActor is used, causing a nonce mismatch.

The fix in this PR constructs CometActor using the new deployment manager's signers when forking CometContext.