Support multiple Comets using one Configurator (#437)
This PR modifies the Configurator implementation to support multiple deployments of Comet so only a single Configurator instance needs to be deployed per chain. Instead of storing a single Configuration and factory address, the Configurator now stores a mapping from Comet proxy addresses to their respective Configuration and factory addresses. Because Configurator.deploy() now takes a Comet proxy address as an argument, a small change is also made to the CometProxyAdmin.deployAndUpgradeTo(...) function to use this new interface. I also removed the logic of setting an initial factory and configuration from the initializer and, instead, added new setters to set these after initialization.

This PR also updates our unit tests, scenarios, deployment logic, and migration scripts to work with the new Configurator implementation.

For scenarios specifically, I added a new flag called upgradeAll to the ModernConstraint that upgrades all contracts. This is needed in scenarios that need to upgrade more than just Comet. For example, we need to upgrade the CometProxyAdmin in a small number of our scenarios that use the Configurator.

TLDR: Contract changes are:

Change Configurator storage for factory and configuration to be mappings
Move the setting of the initial factory and configuration out of the initializer
Modify the CometProxyAdmin to use the new interface for Configurator.deploy(address)