Configurator check for baseToken instead of governor (#450)
As proposed by OpenZeppelin, since `governor` is mutable in configuration. For correctness, we should either do this or update docs that `baseToken` and `trackingIndexScale` are technically mutable. If we check for `governor`, then governance could reset `governor` to `address(0)` and then overwrite the entire config, which would include `baseToken` and `trackingIndexScale`.