Taming migration scripts
We are working towards a simplified/standard workflow which we still need to document:

 1. Open a PR with a migration script
 2. Run 'prepare' workflow in GitHub CI
  - this should deploy any necessary contracts
  - changes to roots should auto-commit to the PR branch
  - maybe go post in the forum about it for review
 3. Run 'enact' workflow in GitHub CI or elsewhere
  - this should make any necessary gov proposals
 4. Once the proposal is accepted by gov, merge PR with new roots

NB: currently `roots.json` changes *can* happen in migrations.
In the next iteration of this, we will have a separate workflow for the (single) 'deploy' script, which should be the only thing modifying roots.
Changes which happen in migrations should be automatically spidered in when any related proposal passes.
Otherwise, a new root should be added and it should be part of 'deploy'.