Migration scripts to upgrade interest rate model on testnet (#477)
This PR adds migration scripts to upgrade the interest rate model on kovan and fuji. The migration script partially re-deploys contracts (only the Configurator proxy, implementation, and CometFactory need to be re-deployed).

I opted for deploying a new Configurator proxy instead of re-using the old one so we can modify the storage layout without having to keep old storage variables around and having to append new storage variables at the end. Rob said this was okay to do since we are still on testnet.

Migrations run:

**Kovan**:
- [x] 1. Prepare migration
 - https://github.com/compound-finance/comet/actions/runs/2693702996
- [x] 2. Enact migration
 - https://github.com/compound-finance/comet/actions/runs/2693862371
- [x] 3. Execute proposal
 - https://kovan.etherscan.io/tx/0x8d858e0b31a99aac89a2081bc4057fc0194a14555be6594c0f14dc872258ed4b

**Fuji**:
- [x] 1. Prepare migration
 - https://github.com/compound-finance/comet/actions/runs/2693704894
- [x] 2. Enact migration
 - https://github.com/compound-finance/comet/actions/runs/2693786849
- [x] 3. Execute proposal
 - https://testnet.snowtrace.io/tx/0x41e189384cba4646b4c9ce86bbbcc39747435c81e0fdcccc72a46b28cfc35e9a