Standardize deploy scripts, and deploy once per base
Cleans up and refactors a bunch of the deploy code.
We no longer have a separate code path for development deploys, and we make them more efficient by only deploying once per base instead of once per scenario.
As a result, we also avoid redeploying all tokens during 'upgrades' in development.

Of note: we store the world in the context to avoid passing it around everywhere, and to prevent incorrect usage.
In addition, the deploy logic now lives in the deployment manager, and is performed by scenario plugin (Worker), which gives the scenario plugin a stronger dependency on the deployment manager dependency.
However these plugins were already coupled for other reasons, and there are likely other benefits that can come of this.